worker_app/requirements.txt
amqp==5.3.1
annotated-types==0.7.0
anthropic==0.54.0
anyio==4.9.0
billiard==4.2.1
cachetools==5.5.2
celery==5.5.3
certifi==2025.6.15
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
-e git+https://github.com/crnorthc/resumai.git@main#egg=common&subdirectory=common
cryptography==45.0.4
distro==1.9.0
google-auth==2.40.3
google-genai==1.20.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
Jinja2==3.1.6
jiter==0.10.0
kombu==5.5.4
lxml==5.4.0
MarkupSafe==3.0.2
openai==1.88.0
packaging==25.0
pillow==11.2.1
prompt_toolkit==3.0.51
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.1.0
redis==6.2.0
reportlab==4.4.1
requests==2.32.4
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.4.0
vine==5.1.0
wcwidth==0.2.13
websockets==15.0.1
